backend/requirements.txt
fastapi>=0.68.0
sqlalchemy>=1.4.23
python-dotenv>=0.19.0
uvicorn>=0.15.0
pydantic>=1.8.2
pymysql>=1.0.2
mysql-connector-python>=8.0.26
httpx>=0.24.0
fastapi-cache2>=0.1.8 